requirements.txt
absl-py==1.1.0
aniso8601==9.0.1
anyascii==0.3.1
astunparse==1.6.3
boto3==1.21.13
botocore==1.24.46
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
filedepot==0.8.0
Flask==2.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.46.3
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.37
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
